+++ InstanceHUB/CoreValuesHUB.cs
@@ -7,13 +7,13 @@ public static class CoreValuesHUB
 
     public static int sampleRate => SampleRate.GetValue();
 
-    public static float melodyPositionInRate => MelodyPositionInRate.GetValue();
+    public static double melodyPositionInRate => MelodyPositionInRate.GetValue();
 
-    public static float melodyLengthInSeconds => 4 / beatsPerSecond;//4 notes per tact
+    public static double melodyLengthInSeconds => 4 / beatsPerSecond;//4 notes per tact
     public static float melodyLengthInUnits => MelodyLengthInUnits.GetValue();
 
     public static int beatsPerMinute => BeatsPerMinute.GetValue();
-    public static float beatsPerSecond => beatsPerMinute / 60f;
+    public static double beatsPerSecond => beatsPerMinute / 60d;
     public static float decibelGate => DecibelGate.GetValue();
     public static int reaction => Reaction.GetValue();
 
@@ -22,7 +22,7 @@ public static class CoreValuesHUB
 
     public static InstanceCell<int> SampleRate = new InstanceCell<int>();
 
-    public static InstanceCell<float> MelodyPositionInRate = new InstanceCell<float>();
+    public static InstanceCell<double> MelodyPositionInRate = new InstanceCell<double>();
 
     public static InstanceCell<float> MelodyLengthInUnits = new InstanceCell<float>();
